@@ -1,81 +1,66 @@
 source 'http://rubygems.org'
 
-gem 'rails', '3.2.16'
-#gem 'rack-cache', '1.2'
+
+gem 'railties'
+gem 'rails', '4.2.7'
+gem 'protected_attributes'
+gem 'responders' #, '~> 2.0' 
 
 group :test, :development do
   gem 'hirb'
 end
 
-gem "lograge"
-gem 'logstash-event'
+gem "lograge" # formatador de log
+gem 'logstash-event' # gerar formato do log compativel com logstash
+# gem 'bunny' #, '~> 1.7.0' # client do rabbitmq
 
+# gem 'tiny_tds', '0.7.0'
+# gem 'activerecord-sqlserver-adapter'
+gem 'will_paginate' #, '~> 3.0'
 
-# gem 'simpleconfig', '~> 2.0', '>= 2.0.1'
-
-gem 'bunny', '~> 1.7.0'
-
-
-gem 'tiny_tds', '0.7.0'
-gem 'activerecord-sqlserver-adapter'
-gem 'will_paginate', '~> 3.0'
-
-# gem 'pg'
 
 #gem 'jira-ruby'
 
-gem 'oauth', '~> 0.4.7' # exigido pelo JIRA, que agora esta dentro do lib
+gem 'oauth' #, '~> 0.4.7' # exigido pelo JIRA, que agora esta dentro do lib
 
-gem 'mysql2', '0.3.20'
-# gem 'activerecord-mysql2-adapter'
+gem 'mysql2' , '0.3.20'
 
-gem 'thin'
 
-gem 'log4r'
+gem 'thin' # web server
+
+# gem 'log4r'
 
 # gem 'foreman'
 
-gem 'nokogiri', '1.6.6.2'
+gem 'nokogiri' , '1.6.6.2' # Nao mudar a versao dessa gem, pois da problema de compilacao (pelo menos no mac)
 
-gem 'axlsx_rails'
+gem 'axlsx_rails' # gerador de Excel via nokogiri
 
-#gem 'pusher'
+# gem 'pusher' # gem de notificacao no browser - pago
 
 # gem 'wisper' # publish/subscriber
 
-# gem 'activeadmin' # TODO - se eu tirar isso, o InheritedResources::Base” SOME !!
-gem 'inherited_resources', '1.6.0'
-# gem 'sass-rails',   '~> 3.2.3'	
-#gem "meta_search",    '>= 1.1.0.pre'   # TODO preciso dessa GEM ??
+gem 'inherited_resources' 
 
-# colocar tag de desenvolvimento TODO
 
-gem 'thinking-sphinx', '2.0.13'
+gem 'thinking-sphinx' #, '2.0.13'
 
 gem 'ruby-graphviz'
 
 gem 'jquery-rails'
 
-gem "audited-activerecord", "~> 3.0"
+# gem "audited-activerecord" #, "~> 3.0"
+gem "audited", "~> 4.3"
 
-gem 'rake', '10.4.2'
+gem 'rake' #, '10.4.2'
 # gem "therubyracer"
 #gem "less-rails"
 #
-gem 'redis-rails',  '~> 3.2'
+gem 'redis-rails' #,  '~> 3.2'
 
-
-
-group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
-
-  gem 'uglifier', '>= 1.0.3'
-  gem 'twitter-bootstrap-rails'
-end
+gem 'sass-rails'
 
 gem "nifty-generators", :group => :development
-
 
 gem "rspec", :group => :test
 
@@ -84,22 +69,17 @@ gem "awesome_print", require:"ap", :group => :development
 
 gem 'omniauth-oauth2', '~> 1.3.1'
 gem 'omniauth-google-oauth2', '0.2.6'
-#gem 'omniauth-office365', github: 'simi/omniauth-office365'
 gem 'cancan' ,  '~> 1.6'
-gem 'dalli', '2.7.4'    #high performance pure Ruby client for accessing memcached servers
+gem 'dalli', '2.7.4'   # cache no redis
 
 
 gem 'whenever', :require => false # crontab like
-gem 'sidekiq', '2.17.8'
-gem 'sidekiq-failures', '0.4.5'
+gem 'sidekiq' 
+gem 'sidekiq-failures' 
 gem 'sinatra', require: false
-# gem 'slim'
 
-gem 'quiet_assets'
+# gem 'quiet_assets' # suppresses output of asset requests 
 
-gem 'deep_cloneable', '~> 1.5.5'
-
-#gem 'paper_trail', '>= 3.0.0.rc2'
-
+gem 'deep_cloneable' 
 
 gem "mocha", :group => :test
