@@ -1,9 +1,3 @@
 # Be sure to restart your server when you modify this file.
 
-# ItMgmt::Application.config.session_store :cookie_store, :key => '_itmgmt_session'
-ItMgmt::Application.config.session_store :redis_store #, 'redis://10.0.2.2:6379/0/cache'
-
-# Use the database for sessions instead of the cookie-based default,
-# which shouldn't be used to store highly confidential information
-# (create the session table with "rails generate session_migration")
-# ItMgmt::Application.config.session_store :active_record_store
+Rails.application.config.session_store :cookie_store, key: '_ItMgmtRail424_session'
