@@ -17,21 +17,21 @@ ItMgmt::Application.routes.draw do
 
 
 
-  get "Inventario/PorHostname/:id", to: "Inventario#porHostname", as: "InventarioPorHostname"
+  get "Inventario/PorHostname/:id", to: "inventario#porHostname", as: "InventarioPorHostname"
   
-  get "Licencas/PorGestor/:id(/:versao)", to: "Licencas#porGestor", as: "LicencasPorGestor" , :constraints => { :id => /[a-zA-z0-9.]*/ }
+  get "Licencas/PorGestor/:id(/:versao)", to: "licencas#porGestor", as: "LicencasPorGestor" , :constraints => { :id => /[a-zA-z0-9.]*/ }
 
-  get "Licencas/PorUsuario/:id", to: "Licencas#porUsuario", as: "LicencasPorUsuario" , :constraints => { :id => /.*/ }
+  get "Licencas/PorUsuario/:id", to: "licencas#porUsuario", as: "LicencasPorUsuario" , :constraints => { :id => /.*/ }
 
 
-  get "identities/:id", to: "Identities#show", as: "identities" , :constraints => { :id => /.*/ }
+  get "identities/:id", to: "identities#show", as: "identities" , :constraints => { :id => /.*/ }
 
   post "ws_register_desligamento", to: "log_desligamento#ws_register_desligamento"
   resources :users
 
 
   resources :schedulers
-  get "schedulers/:id/run", to: "schedulers#run", as: "scheduler_run"
+  # get "schedulers/:id/run", to: "schedulers#run", as: "scheduler_run"
   post "schedulers/:id/run", to: "schedulers#run", as: "scheduler_run"
 
 
@@ -40,27 +40,27 @@ ItMgmt::Application.routes.draw do
 
 
 
-  get "painel_producao/:id/:scope" , to:"painel_producao#index", as: "painel_producao"
-  get "painel_producao/:id" , to:"painel_producao#index", as: "painel_producao"
+  get "painel_producao/:id/(:scope)" , to:"painel_producao#index", as: "painel_producao"
+  # get "painel_producao/:id" , to:"painel_producao#index", as: "painel_producao"
 
 
 
-  match "GestaoUsuarios", to: "GestaoUsuarios#index", as: "gestao_usuarios", :constraints => { :id => /.*/ }
+ get "GestaoUsuarios", to: "gestao_usuarios#index", as: "gestao_usuarios", :constraints => { :id => /.*/ }
 
-  get 'GestaoUsuarios/:id/email', to: "GestaoUsuarios#email", :constraints => { :id => /.*/ },  :as => :email_gestao_usuario
-  post 'GestaoUsuarios/:id/email', to: "GestaoUsuarios#enviar_email", :constraints => { :id => /.*/ },  :as => :enviar_email_gestao_usuario
+  get 'GestaoUsuarios/:id/email', to: "gestao_usuarios#email", :constraints => { :id => /.*/ },  :as => :email_gestao_usuario
+  post 'GestaoUsuarios/:id/email', to: "gestao_usuarios#enviar_email", :constraints => { :id => /.*/ },  :as => :enviar_email_gestao_usuario
 
 
-  post "GestaoUsuarios/AlocarEstacao",  to: "GestaoUsuarios#alocar_estacao", as: "alocar_estacao"
-  get "GestaoUsuarios/EscolherLicencaAlocar",  to: "GestaoUsuarios#escolher_licenca_alocar", as: "escolher_licenca_alocar"
-  get "GestaoUsuarios/EscolherEstacaoAlocar",  to: "GestaoUsuarios#escolher_estacao_alocar", as: "escolher_estacao_alocar"
-  get "GestaoUsuarios/:id/confirmardesalocarestacao/:estacao", to: "GestaoUsuarios#confirmar_desalocar_estacao", as: "confirmar_desalocar_estacao", :constraints => { :id => /.*/ }
-  get "GestaoUsuarios/:id/imprimirtermoentrega/:tipo/:ativo", to: "GestaoUsuarios#imprimir_termo_entrega", as: "imprimir_termo_entrega", :constraints => { :id => /.*/ }
-  get "GestaoUsuarios/:id/imprimirtermodevolucao/:tipo/:ativo", to: "GestaoUsuarios#imprimir_termo_devolucao", as: "imprimir_termo_devolucao", :constraints => { :id => /.*/ }
-  get "GestaoUsuarios/:id/confirmarremocaolicenca/:licenca", to: "GestaoUsuarios#confirmar_remocao_licenca", as: "confirmar_remocao_licenca", :constraints => { :id => /.*/ }
-  post "GestaoUsuarios/AlocarLicenca", to: "GestaoUsuarios#alocar_licenca", as: "alocar_licenca"
-  post "GestaoUsuarios/deletelicenca", to: "GestaoUsuarios#remover_licenca", as: "remover_licenca"
-  post "GestaoUsuarios/DesalocarEstacao", to: "GestaoUsuarios#desalocar_estacao", as: "desalocar_estacao"
+  post "GestaoUsuarios/AlocarEstacao",  to: "gestao_usuarios#alocar_estacao", as: "alocar_estacao"
+  get "GestaoUsuarios/EscolherLicencaAlocar",  to: "gestao_usuarios#escolher_licenca_alocar", as: "escolher_licenca_alocar"
+  get "GestaoUsuarios/EscolherEstacaoAlocar",  to: "gestao_usuarios#escolher_estacao_alocar", as: "escolher_estacao_alocar"
+  get "GestaoUsuarios/:id/confirmardesalocarestacao/:estacao", to: "gestao_usuarios#confirmar_desalocar_estacao", as: "confirmar_desalocar_estacao", :constraints => { :id => /.*/ }
+  get "GestaoUsuarios/:id/imprimirtermoentrega/:tipo/:ativo", to: "gestao_usuarios#imprimir_termo_entrega", as: "imprimir_termo_entrega", :constraints => { :id => /.*/ }
+  get "GestaoUsuarios/:id/imprimirtermodevolucao/:tipo/:ativo", to: "gestao_usuarios#imprimir_termo_devolucao", as: "imprimir_termo_devolucao", :constraints => { :id => /.*/ }
+  get "GestaoUsuarios/:id/confirmarremocaolicenca/:licenca", to: "gestao_usuarios#confirmar_remocao_licenca", as: "confirmar_remocao_licenca", :constraints => { :id => /.*/ }
+  post "GestaoUsuarios/AlocarLicenca", to: "gestao_usuarios#alocar_licenca", as: "alocar_licenca"
+  post "GestaoUsuarios/deletelicenca", to: "gestao_usuarios#remover_licenca", as: "remover_licenca"
+  post "GestaoUsuarios/DesalocarEstacao", to: "gestao_usuarios#desalocar_estacao", as: "desalocar_estacao"
 
   # match '/404' => 'errors#not_found'
   # match '/422' => 'errors#server_error'
@@ -75,7 +75,7 @@ ItMgmt::Application.routes.draw do
   resources :cadsurveys
 
 
-  match "survey/:id", to: "survey#show"
+  # match "survey/:id", to: "survey#show"
 
   get "ajax/cis"
   get "ajax/gestores"
@@ -88,13 +88,11 @@ ItMgmt::Application.routes.draw do
 
   resources :events, :only => [:index]
 
-  resources :fornecedores
-
   resources :mailings
-  match "mailings/enviar_email", to: "mailings#enviar_email"
-  match "mailings/enviar_email_sql", to: "mailings#enviar_email_sql"
-  match "mailings/confirma_enviar_email", to: "mailings#confirma_enviar_email"
-  match "mailings/confirma_enviar_email_sql", to: "mailings#confirma_enviar_email_sql"
+  get  "mailings/enviar_email", to: "mailings#enviar_email"
+  get "mailings/enviar_email_sql", to: "mailings#enviar_email_sql"
+  get "mailings/confirma_enviar_email", to: "mailings#confirma_enviar_email"
+  get "mailings/confirma_enviar_email_sql", to: "mailings#confirma_enviar_email_sql"
 
 
 
@@ -121,8 +119,6 @@ ItMgmt::Application.routes.draw do
   resources :snippets
   #match 'mailer(/:action(/:id(.:format)))' => 'mailer#:action'
 
-  resources :exec_itens_checklists
-
   resources :ramal, :only => [:index]
   resources :exec_checklists, :only => [:show, :index]
 
@@ -132,52 +128,35 @@ ItMgmt::Application.routes.draw do
   end
 
 
-  resources :itens_checklists
-
-
-  #resources :status_checklists
-
-
   resources :audits
-
-
-  #match "audits", to: "audits#index"
-  #match "/audits/:id", to: "audits#show"
-
-  #mount Sidekiq::Web, at: '/sidekiq'
-  # match "ramais", to: "ramal_login#index"
-  # match "ramais/new", to: "ramal_login#new"
-  # match "tel_ramal_login", to: "ramal_login#index"
-  # match "tel_ramal_logins", to: "ramal_login#index" #TODO limpar esse lixo..
-
   resources :tipotasks
   resources :tipo_tasks
   resources :atributos
 
 
-  match 'auth/:provider/callback', to: 'sessions#create'
-  match 'auth/failure', to: redirect('/')
-  match 'signout', to: 'sessions#destroy', as: 'signout'
+  match 'auth/:provider/callback', to: 'sessions#create', via: [:get, :post]
+  match 'auth/failure', to: redirect('/'), via: [:get, :post]
+  match 'signout', to: 'sessions#destroy', as: 'signout', via: [:get, :post]
 
-  match 'destroy_admin_user_session_path', to: 'sessions#destroy'
-  match 'destroy_admin_user_session', to: 'sessions#destroy'
+  match 'destroy_admin_user_session_path', to: 'sessions#destroy', via: [:get, :post]
+  match 'destroy_admin_user_session', to: 'sessions#destroy', via: [:get, :post]
 
   resources :cadrelatorios
-  match 'cadrelatorios/:id/duplicar', to: "cadrelatorios#duplicar", :as => :duplicar_relatorio
+  match 'cadrelatorios/:id/duplicar', to: "cadrelatorios#duplicar", :as => :duplicar_relatorio, via: [:get, :post]
 
   resources :statuscis
 
   # para todos nao ser confundido com ID
 
 
-  match 'relatorio/:id', to: "relatorio#index", as: "relatorio"
-  match 'email/:acao/:id', to: "email#enviar"
-  match 'email/enviar_email/:id', to: "email#enviar+email"
+  match 'relatorio/:id', to: "relatorio#index", as: "relatorio", via: [:get, :post]
+  match 'email/:acao/:id', to: "email#enviar", via: [:get, :post]
+  match 'email/enviar_email/:id', to: "email#enviar+email", via: [:get, :post]
 
-  match '/tasks/todos', to: 'tasks#todos'
-  match '/cis/todos', to: 'cis#todos'
+  match '/tasks/todos', to: 'tasks#todos', via: [:get, :post]
+  match '/cis/todos', to: 'cis#todos', via: [:get, :post]
 
-  match '/tasks/:id/new_from_ci', to: 'tasks#new_from_ci', as: "tasks_new_from_ci"
+  match '/tasks/:id/new_from_ci', to: 'tasks#new_from_ci', as: "tasks_new_from_ci", via: [:get, :post]
 
   resources :parametros #, :only => [:index,:edit,:update,:new]
 
@@ -197,38 +176,38 @@ ItMgmt::Application.routes.draw do
   get "cis/:id/log", to: "cis#log", as: "log_ci"
   post 'cis/:id/log', to: "cis#register_log",  :as => :register_log
 
-  match 'cis/:id/impactados', to: "cis#impactados",  :as => :impactados
-  match 'cis/:id/dependentes', to: "cis#dependentes",  :as => :dependentes
-  match 'cis/:id/dependentes_all', to: "cis#dependentes_all",  :as => :dependentes_all
-  match 'cis/:id/grafico_impactados', to: "cis#grafico_impactados",  :as => :grafico_impactados
-  match 'cis/:id/impactados_h', to: "cis#impactados_h", :as => :impactados_h
-  match 'cis/:id/dependentes_h', to: "cis#dependentes_h", :as => :dependentes_h
-  #match 'cis/:id/abrir_alerta', to: "cis#abrir_alerta", :as => :abrir_alerta
+  get 'cis/:id/impactados', to: "cis#impactados",  :as => :impactados
+  get 'cis/:id/dependentes', to: "cis#dependentes",  :as => :dependentes
+  get 'cis/:id/dependentes_all', to: "cis#dependentes_all",  :as => :dependentes_all
+  get 'cis/:id/grafico_impactados', to: "cis#grafico_impactados",  :as => :grafico_impactados
+  get 'cis/:id/impactados_h', to: "cis#impactados_h", :as => :impactados_h
+  get 'cis/:id/dependentes_h', to: "cis#dependentes_h", :as => :dependentes_h
+  #get 'cis/:id/abrir_alerta', to: "cis#abrir_alerta", :as => :abrir_alerta
 
   post 'cis/:id/novo_dependente', to: "cis#gera_novo_dependente", :as => :gera_novo_dependente
-  match 'cis/:id/novo_dependente', to: "cis#novo_dependente", :as => :novo_dependente
+  match 'cis/:id/novo_dependente', to: "cis#novo_dependente", :as => :novo_dependente, via: [:get, :post]
   post 'cis/:id/novo_impactado', to: "cis#gera_novo_impactado", :as => :gera_novo_impactado
-  match 'cis/:id/novo_impactado', to: "cis#novo_impactado", :as => :novo_impactado
+  match 'cis/:id/novo_impactado', to: "cis#novo_impactado", :as => :novo_impactado, via: [:get, :post]
 
   post 'cis/:id/elimina_dependente', to: "cis#elimina_dependente", :as => :elimina_dependente
-  match 'cis/:id/elimina_dependente', to: "cis#ask_elimina_dependente", :as => :ask_elimina_dependente
+  match 'cis/:id/elimina_dependente', to: "cis#ask_elimina_dependente", :as => :ask_elimina_dependente, via: [:get, :post]
   post 'cis/:id/elimina_impactado', to: "cis#elimina_impactado", :as => :elimina_impactado
-  match 'cis/:id/elimina_impactado', to: "cis#ask_elimina_impactado", :as => :ask_elimina_impactado
+  match 'cis/:id/elimina_impactado', to: "cis#ask_elimina_impactado", :as => :ask_elimina_impactado, via: [:get, :post]
 
   post 'cis/:id/duplicar_ci', to: "cis#duplicar_ci", :as => :duplicar_ci
-  match 'cis/:id/duplicar_ci', to: "cis#ask_duplicar_ci", :as => :ask_duplicar_ci
+  match 'cis/:id/duplicar_ci', to: "cis#ask_duplicar_ci", :as => :ask_duplicar_ci, via: [:get, :post]
   
 
-  match 'cis/:id/confirmar_eliminacao', to: "cis#confirmar_eliminacao", :as => :confirmar_eliminacao
+  match 'cis/:id/confirmar_eliminacao', to: "cis#confirmar_eliminacao", :as => :confirmar_eliminacao, via: [:get, :post]
   post 'cis/:id/cis_eliminar', to: "cis#eliminar", :as => :eliminar_ci
-  match 'cis/checkChave', to: "cis#check_chave"
+  match 'cis/checkChave', to: "cis#check_chave", via: [:get, :post]
 
-  match 'cis/massiveupdate', to: "cis#massiveupdate"
-  match '/cis2', to: 'cis#index2'
+  match 'cis/massiveupdate', to: "cis#massiveupdate", via: [:get, :post]
+  # match '/cis2', to: 'cis#index2'
 
 
-  match 'cis/search', to: "cis#search"
-  match 'CMDB', to: "cis#index"
+  get 'cis/search', to: "cis#search"
+  get 'CMDB', to: "cis#index"
 
   resources :cis
   resources :criticidades
@@ -245,11 +224,11 @@ ItMgmt::Application.routes.draw do
 
   resources :tipocis
 
-  match '/404' => 'errors#not_found'            , :via => :all
-  match '/422' => 'errors#server_error'         , :via => :all
-  match '/500' => 'errors#internal_server_error', :via => :all
+  get '/404' => 'errors#not_found'            , :via => :all
+  get '/422' => 'errors#server_error'         , :via => :all
+  get '/500' => 'errors#internal_server_error', :via => :all
 
-  match '*a', :to => 'cis#index'
+  get '*a', :to => 'cis#index'
 
 
 end
