@@ -1,20 +1,22 @@
 class PainelProducaoController < ApplicationController
+
   def index
-  	@painel = []
-  	Cadrelatorio.where("dashboard like \"%#{params[:id]}%\"").each do |c|
-    _scope = params[:scope] || "ALL"
-    # Cadrelatorio.all.each do |c|
-		begin
-            mysql_res = ActiveRecord::Base.connection.execute("SET SESSION group_concat_max_len = 10000;")
-            sql_res = ActiveRecord::Base.connection.execute(c.consulta)
-            puts _scope
-            puts sql_res.count
-            puts (_scope=="ALL") || (sql_res.count>0)
-            @painel << [c.nome,c.descricao,sql_res.count] if (_scope=="ALL") || (sql_res.count>0)
-        rescue
-			 @painel << [c.nome,c.descricao,"Consulta com Erro"]
-	  	end
-	end
+    @painel = []
+    Cadrelatorio.where("dashboard like \"%#{params[:id]}%\"").each do |c|
+      _scope = params[:scope] || "ALL"
+      Rails.logger.debug "[DEBUG]PainelProducaoController:index params:#{params} - scope: #{_scope}"
+      @nomePainel = params[:id]
+      # Cadrelatorio.all.each do |c|
+      begin
+        # Rails.logger.debug "[DEBUG]PainelProducaoController:index query=#{c}"
+        mysql_res = ActiveRecord::Base.connection.execute("SET SESSION group_concat_max_len = 10000;")
+        sql_res = ActiveRecord::Base.connection.execute(c.consulta)
+        # Rails.logger.debug "[DEBUG]PainelProducaoController:index count=#{sql_res.count}"
+        @painel << [c.nome,c.descricao,sql_res.count] if (_scope=="ALL") || (sql_res.count>0)
+      rescue
+        @painel << [c.nome,c.descricao,"Consulta com Erro"]
+      end
+    end
   end
 
 end
