@@ -1,6 +1,6 @@
 class SchedulersController < InheritedResources::Base
 
-	load_and_authorize_resource
+load_and_authorize_resource
 
 def run
 	id = params[:id]
