@@ -6,7 +6,7 @@ class Software < ActiveRecord::Base
             :class_name => "InventarioSw",
             :foreign_key => "software",
             :primary_key => "software"
-  default_scope order('software ASC')
+  default_scope { order('software ASC') }
 
   def self.licencasExistentes
   	self.all.map{|s| {software: s.software,custoMensal: s.custoMensal,status: s.status}}
