@@ -2,12 +2,14 @@ require File.expand_path('../boot', __FILE__)
 
 require 'rails/all'
 
-if defined?(Bundler)
-  # If you precompile assets before deploying to production, use this line
-  Bundler.require(*Rails.groups(:assets => %w(development test)))
-  # If you want your assets lazily compiled in production, use this line
-  # Bundler.require(:default, :assets, Rails.env)
-end
+# if defined?(Bundler)
+#   # If you precompile assets before deploying to production, use this line
+#   Bundler.require(*Rails.groups(:assets => %w(development test)))
+#   # If you want your assets lazily compiled in production, use this line
+#   # Bundler.require(:default, :assets, Rails.env)
+# end
+
+Bundler.require(*Rails.groups)
 
 module ItMgmt 
   class Application < Rails::Application
@@ -19,7 +21,6 @@ module ItMgmt
     # config.autoload_paths += %W(
     #     #{config.root}/app/controllers/custom
     # )
-
 
     config.cache_store = :redis_store #, 'redis://10.0.2.2:6379/0/cache', { expires_in: 90.minutes }
     # config.cache_classes = true
@@ -47,7 +48,7 @@ module ItMgmt
     config.encoding = "utf-8"
 
     # Configure sensitive parameters which will be filtered from the log file.
-    config.filter_parameters += [:password]
+    # config.filter_parameters += [:password]
 
     # Enable escaping HTML in JSON.
     config.active_support.escape_html_entities_in_json = true
@@ -61,13 +62,8 @@ module ItMgmt
     # This will create an empty whitelist of attributes available for mass-assignment for all models
     # in your app. As such, your models will need to explicitly whitelist or blacklist accessible
     # parameters by using an attr_accessible or attr_protected declaration.
-    config.active_record.whitelist_attributes = false
+    # config.active_record.whitelist_attributes = false
 
-    # Enable the asset pipeline
-    config.assets.enabled = true
-
-    # Version of your assets, change this if you want to expire all your assets
-    config.assets.version = '1.0'
 
     config.exceptions_app = self.routes
     
@@ -82,6 +78,7 @@ module ItMgmt
     config.after_initialize do
        Dir["./app/rules/*.rb"].each {|file| require file }
     end
+
   end
 end
 
